Changed by-zone usage graph to Gallons instead of %, also fixed few bugs
Changed by-zone usage graph to Gallons instead of %, also fixed few bugs
in by-zone usage visualization calculation and in WWCounters update
(counters on the Home page).

+++ Station/sdlog.h
@@ -160,9 +160,9 @@ public:
         bool begin(void);
         void Close();
         // Watering activity logging. Note: signature is deliberately compatible with sprinklers_pi control program
-        bool LogZoneEvent(time_t start, int zone, int duration, int water_used, int schedule, int sadj, int wunderground);
+        bool LogZoneEvent(time_t start, int zone, int duration, uint16_t water_used, int schedule, int sadj, int wunderground);
 		// Log whole schedule event
-		bool LogSchedEvent(time_t start, int duration, int water_used, int schedule, int sadj, int wunderground);
+		bool LogSchedEvent(time_t start, int duration, uint16_t water_used, int schedule, int sadj, int wunderground);
 
         // Emit zone log watering data 
         bool TableZone(FILE* stream_file, time_t start, time_t end);
